Use sequence type in the config definition
