+++ src/widgets/immunizations/immunizations-detailed-summary.css
@@ -1,3 +1,11 @@
+:root {
+  --omrs-bg-immunization-sequence-odd-row: #c4d7e6;
+  --omrs-bg-immunization-sequence-even-row: #e5f1fb;
+  --omrs-bg-immunization-odd-row: #fff;
+  --omrs-bg-immunization-even-row: #eee;
+  --omrs-bg-immunization-header: #c8cdd2;
+}
+
 :global(.omrs-breakpoint-gt-tablet) .immunizationSummary {
   width: 45.5vw;
 }
